api: make the NCO-disable path of SetNCOIndex reachable
SetNCOIndex took the index as uint8_t, so the index < 0 branches that
bypass the CMIX could never be taken - the disable sentinel implied by
the int-returning GetNCOIndex (and used by the legacy LMS_SetNCOIndex)
silently wrapped to 255 and was rejected. Take a signed index, validate
the range before any register write, and skip the table selection when
disabling.

Fixes #295

+++ src/boards/MMX8/MM_X8.cpp
@@ -222,7 +222,7 @@ int LimeSDR_MMX8::GetNCOIndex(uint8_t moduleIndex, TRXDir trx, uint8_t channel)
     return mSubDevices[moduleIndex]->GetNCOIndex(0, trx, channel);
 }
 
-OpStatus LimeSDR_MMX8::SetNCOIndex(uint8_t moduleIndex, TRXDir trx, uint8_t channel, uint8_t index, bool downconv)
+OpStatus LimeSDR_MMX8::SetNCOIndex(uint8_t moduleIndex, TRXDir trx, uint8_t channel, int16_t index, bool downconv)
 {
     if (moduleIndex >= mSubDevices.size())
         return OpStatus::OutOfRange;
